only delete /buildlogs/build_core/faileddownloads if it exists
git-svn-id: https://svn.code.sf.net/p/rebeccablackos/code@5383 b52b6941-3400-464c-9f42-43200397181c

+++ externalbuilders/rebeccablackos_phase1.sh
@@ -89,7 +89,10 @@ mount --bind "$BUILDLOCATION"/build/"$BUILDARCH"/archives "$BUILDLOCATION"/build
 mount --bind "$BUILDLOCATION"/build/"$BUILDARCH"/buildlogs "$BUILDLOCATION"/build/"$BUILDARCH"/workdir/buildlogs
 
 #Clear list of failed downloads
-rm "$BUILDLOCATION"/build/"$BUILDARCH"/buildlogs/build_core/faileddownloads
+if [[ -e "$BUILDLOCATION"/build/"$BUILDARCH"/buildlogs/build_core/faileddownloads ]]
+then
+  rm "$BUILDLOCATION"/build/"$BUILDARCH"/buildlogs/build_core/faileddownloads
+fi
 
 #Copy in system resolv.conf
 rm "$BUILDLOCATION"/build/"$BUILDARCH"/workdir/etc/resolv.conf
